Renamed controller to bsmp_sim
